Tryout of working_copy? method
Should indicate whether a document is a working copy, note that this
method is not tested yet for lack of examples. Also note that this
assumes there is no via link for  normal documents.

This could be used in the save/put method to avoid the checkin flag for
documents that do not need it.