Use super when possible
Note that this method probably is wrong anyway: it seems that all
attributes are always present directly, and that it therefore is silly
to call super and wrong not to filter the attributes when inherited is
no.

On the other hand we currently assume that attributes gives all
attributes in a few places.