Change way reload/self_link works
The problem was that we did not have a self_link that we could PUT
through, meaning we could not update. So we return to the previous way
of self_link referencing the @self_link, but we also allow parameters to
it (I haven't checked if this actually works for GET)

Additionally reload throws exception if the object was never saved
before.